agent/oauth_test/requirements.txt
praw>=7.7.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0